restructuring all test to make it readable
