+++ .github/workflows/ds-parquet.yml
@@ -102,7 +102,7 @@ jobs:
         uses: actions/cache@v2.1.5
         with:
           path: ~/.m2
-          key: ${{ runner.os }}-m2-${{ hashFiles('/pom.xml') }}
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
           restore-keys: ${{ runner.os }}-m2
       - name: Run ZooKeeper
         run: |
