+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/CombinedInsertTests.java
@@ -36,16 +36,10 @@ public class CombinedInsertTests {
     private static final long NON_ALIGNED_ROW_START_TIMESTAMP = 30001L;
     private static final long NON_ALIGNED_ROW_END_TIMESTAMP = 40000L;
 
-    private static Session session = null;
+    private static MultiConnection session = null;
 
-    private static SessionPool sessionPool = null;
-
-    public CombinedInsertTests(Session passedSession){
+    public CombinedInsertTests(MultiConnection passedSession){
         session = passedSession;
-    }
-
-    public CombinedInsertTests(SessionPool passedSessionPool){
-        sessionPool = passedSessionPool;
     }
 
     public void testInserts() throws SessionException, ExecutionException {
@@ -95,10 +89,7 @@ public class CombinedInsertTests {
         }
 
         System.out.println("insertColumnRecords...");
-        if(session != null)
-            session.insertColumnRecords(paths, timestamps, valuesList, dataTypeList, null);
-        else if(sessionPool != null)
-            sessionPool.insertColumnRecords(paths, timestamps, valuesList, dataTypeList, null);
+        session.insertColumnRecords(paths, timestamps, valuesList, dataTypeList, null);
     }
 
     private static void insertNonAlignedColumnRecords() throws SessionException, ExecutionException {
@@ -141,10 +132,7 @@ public class CombinedInsertTests {
         }
 
         System.out.println("insertNonAlignedColumnRecords...");
-        if(session != null)
-            session.insertNonAlignedColumnRecords(paths, timestamps, valuesList, dataTypeList, null);
-        else if(sessionPool != null)
-            sessionPool.insertNonAlignedColumnRecords(paths, timestamps, valuesList, dataTypeList, null);
+        session.insertNonAlignedColumnRecords(paths, timestamps, valuesList, dataTypeList, null);
     }
 
     private static void insertRowRecords() throws SessionException, ExecutionException {
@@ -182,10 +170,7 @@ public class CombinedInsertTests {
         }
 
         System.out.println("insertRowRecords...");
-        if(session != null)
-            session.insertRowRecords(paths, timestamps, valuesList, dataTypeList, null);
-        else if(sessionPool != null)
-            sessionPool.insertRowRecords(paths, timestamps, valuesList, dataTypeList, null);
+        session.insertRowRecords(paths, timestamps, valuesList, dataTypeList, null);
     }
 
     private static void insertNonAlignedRowRecords() throws SessionException, ExecutionException {
@@ -226,10 +211,7 @@ public class CombinedInsertTests {
         }
 
         System.out.println("insertNonAlignedRowRecords...");
-        if(session != null)
-            session.insertNonAlignedRowRecords(paths, timestamps, valuesList, dataTypeList, null);
-        else if(sessionPool != null)
-            sessionPool.insertNonAlignedRowRecords(paths, timestamps, valuesList, dataTypeList, null);
+        session.insertNonAlignedRowRecords(paths, timestamps, valuesList, dataTypeList, null);
     }
 
 }
