--- test/src/test/java/cn/edu/tsinghua/iginx/integration/TransformIT.java
@@ -1,5 +1,0 @@
-package cn.edu.tsinghua.iginx.integration;
-
-public class TransformIT {
-
-}
