+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/influxdb/InfluxDBSQLSessionIT.java
@@ -1,4 +1,6 @@
-package cn.edu.tsinghua.iginx.integration;
+package cn.edu.tsinghua.iginx.integration.influxdb;
+
+import cn.edu.tsinghua.iginx.integration.SQLSessionIT;
 
 public class InfluxDBSQLSessionIT extends SQLSessionIT {
     public InfluxDBSQLSessionIT() {
