+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/influxdb/InfluxDBSQLSessionPoolIT.java
@@ -1,4 +1,6 @@
-package cn.edu.tsinghua.iginx.integration;
+package cn.edu.tsinghua.iginx.integration.influxdb;
+
+import cn.edu.tsinghua.iginx.integration.SQLSessionPoolIT;
 
 public class InfluxDBSQLSessionPoolIT extends SQLSessionPoolIT {
     public InfluxDBSQLSessionPoolIT() {
