+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/iotdb/IoTDB12SessionIT.java
@@ -1,17 +1,20 @@
-package cn.edu.tsinghua.iginx.integration;
+package cn.edu.tsinghua.iginx.integration.iotdb;
+
+import cn.edu.tsinghua.iginx.integration.BaseSessionIT;
 
 import java.util.LinkedHashMap;
 
-public class IoTDB11SessionPoolIT extends BaseSessionPoolIT {
+public class IoTDB12SessionIT extends BaseSessionIT {
 
-    public IoTDB11SessionPoolIT() {
+    public IoTDB12SessionIT() {
         super();
         this.defaultPort2 = 6668;
         this.isAbleToDelete = true;
-        this.storageEngineType = "iotdb11";
+        this.storageEngineType = "iotdb12";
         this.extraParams = new LinkedHashMap<>();
         this.extraParams.put("username", "root");
         this.extraParams.put("password", "root");
         this.extraParams.put("sessionPoolSize", "100");
     }
 }
+
