+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/iotdb/IoTDBSQLSessionIT.java
@@ -1,4 +1,6 @@
-package cn.edu.tsinghua.iginx.integration;
+package cn.edu.tsinghua.iginx.integration.iotdb;
+
+import cn.edu.tsinghua.iginx.integration.SQLSessionIT;
 
 public class IoTDBSQLSessionIT extends SQLSessionIT {
     public IoTDBSQLSessionIT() {
