+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/udf/TransformIT.java
@@ -1,0 +1,5 @@
+package cn.edu.tsinghua.iginx.integration.udf;
+
+public class TransformIT {
+
+}
