+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/udf/UDFIT.java
@@ -1,4 +1,4 @@
-package cn.edu.tsinghua.iginx.integration;
+package cn.edu.tsinghua.iginx.integration.udf;
 
 import cn.edu.tsinghua.iginx.exceptions.ExecutionException;
 import cn.edu.tsinghua.iginx.exceptions.SessionException;
